Fix crasher
Crasher: bug in note trigger logic could switch to a new instrument that is shorter than
the current play position in the current instrument, causing a panic reading off the end of
the shorter sample data. Fix was to only assign the instrument if there is a note (there wasn't
in this case)

Looping sample sanitization: Ignore loop lengths < 4 bytes.