Removed song ended channel
The delays receiving messages on the "song ended" channel led to
unreliable playback in the client code. Player now has a
Playing/Stopped state and Start() and Stop() methods. The player
will transition to the Stopped state when the end of the song is
reached.

GenerateAudio() now returns the number of stereo samples generated.