Attempted fix for notes w/o instruments and vice versa
Sometimes MODs will specify the instrument on a separate row before a
note without an instrument. This broke an assumption in the player that
every note had an instrument, causing the wrong instruments to play, or
nothing to play at all.

This fix has fixed the most obvious errors but there may still be
lingering issues.