Bug fixes to CombFixed
The delay offset was implemented at half of what it should have been.
Instead of a 1s reverb delay it would be 0.5s. This happens because the
computed delay offset index did not account for sample pairs.

When a reverb application wraps the end of the buffer the second part
used an incorrect offset about where to write reverb data.