A rewrite of the note trigger logic
Getting tired of the whack-a-mole bugs: track down and fix a note
triggering bug in one song, only to create a bug in another song. So
doing some testing work in ST3 to figure out the note triggering logic
and then implementing that here. Hoping to also add some unit tests of
the logic.