Fix a gnarly first tick bug in the player
Discovered through unit tests that player initialization and tick
handling logic were incorrect. This caused only a single tick to be
processed for the first row before the player advanced to the next row.
Tick handling proceeded correctly for the remainder of the song. There
are now several TODOs in the code to determine how to handle certain
effects and edge cases w.r.t. tick processing.

Subsequently, working on a unit test that advanced using GenerateAudio
exposed a similar logic bug in that code which will need a separate
commit to fully fix. The advantage of using GenerateAudio to advance is
that we can see sample positions being updated which allows us to verify
that samples are being stopped.