GenerateAudio now generates audio immediately
Previously the first call to GenerateAudio would return a buffer full of
empty audio at the start of the song. This was because the counters used
in GenerateAudio were not initialized correctly (fixed in prior commit)
and the player logic was only ticked after audio generation. Now this
happens before audio generation.

As a consequence, if the player generates exactly the full amount of
audio samples for a tick it will not advance the player to the new tick
until the next call to GenerateAudio.