Add support for S3M's Porta+Vol Slide effect
Introduces a SongType field to the Song which used by the player to
support differences in Porta+Vol Slide effect processing between MOD and
S3M.

Also changes S3M volume slide behavior to give the down parameter
precedence if both down and up are set. This was verified by
experimentation in ST3.